chore/#23 : 컴포넌트 패키지 위치 변경
