+++ app/src/main/java/com/depromeet/team6/presentation/ui/searchlocation/component/SearchLocationTextField.kt
@@ -1,4 +1,4 @@
-package com.depromeet.team6.presentation.ui.home.component
+package com.depromeet.team6.presentation.ui.searchlocation.component
 
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -23,7 +23,7 @@ import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.depromeet.team6.R
-import com.depromeet.team6.presentation.ui.home.component.SearchLocationTextField.MAX_LENGTH
+import com.depromeet.team6.presentation.ui.searchlocation.component.SearchLocationTextField.MAX_LENGTH
 import com.depromeet.team6.presentation.util.modifier.noRippleClickable
 import com.depromeet.team6.presentation.util.modifier.roundedBackgroundWithPadding
 import com.depromeet.team6.ui.theme.defaultTeam6Colors
@@ -53,7 +53,7 @@ fun SearchLocationTextField(
             modifier = modifier
                 .weight(1f)
                 .roundedBackgroundWithPadding(
-                    backgroundColor = defaultTeam6Colors.greyElevatedBackground,
+                    backgroundColor = defaultTeam6Colors.greyQuaternaryLabel,
                     cornerRadius = 8.dp
                 )
                 .padding(horizontal = 12.dp),
@@ -83,14 +83,14 @@ fun SearchLocationTextField(
                 keyboardActions = keyboardActions,
                 keyboardOptions = keyboardOptions,
                 visualTransformation = visualTransformation,
-                textStyle = defaultTeam6Typography.bodyRegular15.copy(color = defaultTeam6Colors.white),
+                textStyle = defaultTeam6Typography.bodyRegular17.copy(color = defaultTeam6Colors.white),
                 decorationBox = { innerTextField ->
                     innerTextField()
                     if (value.isEmpty()) {
                         Text(
                             text = stringResource(R.string.home_search_location_search_text),
-                            color = defaultTeam6Colors.greyQuaternaryLabel,
-                            style = defaultTeam6Typography.bodyMedium15
+                            color = defaultTeam6Colors.greySecondaryLabel,
+                            style = defaultTeam6Typography.bodyRegular17
                         )
                     }
                 }
